Moved ProjectStatus enum to separate file
