+++ src/angular/hq/src/app/charge-code/services/ChargeCodeListService.ts
@@ -1,8 +1,8 @@
 import { Injectable } from '@angular/core';
 import { FormControl } from '@angular/forms';
 import { BehaviorSubject } from 'rxjs';
-import { ProjectStatus } from '../../clients/client-details.service';
 import { GetPSRTimeRecordStaffV1 } from '../../models/PSR/get-psr-time-v1';
+import { ProjectStatus } from '../../enums/project-status';
 
 export enum ActivityName {
   Support = 0,
