+++ src/angular/hq/src/app/enums/project-status.ts
@@ -1,0 +1,12 @@
+export enum ProjectStatus {
+  Unknown = 0,
+  Draft = 1,
+  WaitingForSale = 2,
+  WaitingForClient = 3,
+  WaitingForStaff = 4,
+  InProduction = 5,
+  Ongoing = 6,
+  Completed = 7,
+  Closed = 8,
+  Lost = 9,
+}
