+++ src/angular/hq/src/app/models/projects/get-project-v1.ts
@@ -1,4 +1,4 @@
-import { ProjectStatus } from '../../clients/client-details.service';
+import { ProjectStatus } from '../../enums/project-status';
 import { Period } from '../../projects/project-create/project-create.component';
 import { PagedRequestV1 } from '../common/paged-request-v1';
 import { PagedResponseV1 } from '../common/paged-response-v1';
