+++ src/angular/hq/src/app/projects/project-list/project-list.component.ts
@@ -12,7 +12,6 @@ import {
   switchMap,
   shareReplay,
 } from 'rxjs';
-import { ProjectStatus } from '../../clients/client-details.service';
 import { SortColumn } from '../../models/projects/get-project-v1';
 import { SortDirection } from '../../models/common/sort-direction';
 import { HQService } from '../../services/hq.service';
@@ -26,6 +25,7 @@ import { InRolePipe } from '../../pipes/in-role.pipe';
 import { ProjectListSearchFilterComponent } from '../project-list-search-filter/project-list-search-filter.component';
 import { ProjectSearchFilterService } from '../services/ProjectSearchFilterService';
 import { Period } from '../../models/times/get-time-v1';
+import { ProjectStatus } from '../../enums/project-status';
 
 @Component({
   selector: 'hq-project-list',
