+++ src/angular/hq/src/app/projects/services/ProjectSearchFilterService.ts
@@ -3,8 +3,8 @@ import { FormControl } from '@angular/forms';
 import { BehaviorSubject, first, map } from 'rxjs';
 
 import { HQService } from '../../services/hq.service';
-import { ProjectStatus } from '../../clients/client-details.service';
 import { GetPSRTimeRecordStaffV1 } from '../../models/PSR/get-psr-time-v1';
+import { ProjectStatus } from '../../enums/project-status';
 
 @Injectable({
   providedIn: 'root',
