+++ src/angular/hq/src/app/psr/psr-service.ts
@@ -1,10 +1,10 @@
 import { Injectable, OnDestroy } from '@angular/core';
 import { FormControl } from '@angular/forms';
-import { ProjectStatus } from '../clients/client-details.service';
 import { BehaviorSubject } from 'rxjs';
 import { HQService } from '../services/hq.service';
 import { GetPSRTimeRecordStaffV1 } from '../models/PSR/get-psr-time-v1';
 import { GetProjectActivityRecordV1 } from '../models/PSR/get-project-activity-v1';
+import { ProjectStatus } from '../enums/project-status';
 
 export enum ActivityName {
   Support = 0,
