+++ src/angular/hq/src/app/psr/psrlist/psrlist.component.ts
@@ -14,7 +14,6 @@ import {
   shareReplay,
   firstValueFrom,
 } from 'rxjs';
-import { ProjectStatus } from '../../clients/client-details.service';
 import { GetPSRRecordV1, SortColumn } from '../../models/PSR/get-PSR-v1';
 import { SortDirection } from '../../models/common/sort-direction';
 import { HQService } from '../../services/hq.service';
@@ -24,6 +23,7 @@ import { SortIconComponent } from '../../common/sort-icon/sort-icon.component';
 import { Period } from '../../projects/project-create/project-create.component';
 import { OidcSecurityService } from 'angular-auth-oidc-client';
 import { PsrListService } from './services/pstlistService';
+import { ProjectStatus } from '../../enums/project-status';
 
 @Component({
   selector: 'hq-psrlist',
